Support test user authentication even without authentication provider
